Second Version

+++ app/src/main/java/com/oleh/myweatherapp/weather/Weather.java
@@ -12,12 +12,12 @@ public class Weather implements Serializable {
     private final String weatherDescription;
     private final WeatherIconCode iconCode;
     private final int temp;
-    private final double tempFeelsLike;
-    private final double minTemp;
-    private final double maxTemp;
+    private final int tempFeelsLike;
+    private final int minTemp;
+    private final int maxTemp;
     private final String locationName;
 
-    public Weather(double lat, double lon, String weatherDescription, String iconCode, int temp, double tempFeelsLike, double minTemp, double maxTemp, String locationName) {
+    public Weather(double lat, double lon, String weatherDescription, String iconCode, int temp, int tempFeelsLike, int minTemp, int maxTemp, String locationName) {
         this.lat = lat;
         this.lon = lon;
         this.weatherDescription = weatherDescription;
@@ -49,15 +49,15 @@ public class Weather implements Serializable {
         return temp;
     }
 
-    public double getTempFeelsLike() {
+    public int getTempFeelsLike() {
         return tempFeelsLike;
     }
 
-    public double getMinTemp() {
+    public int getMinTemp() {
         return minTemp;
     }
 
-    public double getMaxTemp() {
+    public int getMaxTemp() {
         return maxTemp;
     }
